requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
altgraph==0.17
apipkg==1.5
Appium-Python-Client==0.50
atomicwrites==1.4.0
attrs==19.3.0
bcrypt==3.1.7
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
cheroot==8.3.0
CherryPy==18.5.0
click==7.1.2
colorama==0.4.3
concurrent-log-handler==0.9.9
cryptography==2.9.2
cx-Oracle==7.3.0
ddt==1.2.2
Django==2.0.4
elasticsearch==7.8.0
execnet==1.7.1
Flask==1.0.2
future==0.18.2
idna==2.8
importlib-metadata==1.6.0
importlib-resources==1.4.0
iniconfig==1.0.1
itsdangerous==1.1.0
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.0
jaraco.text==3.2.0
Jinja2==2.10
jsonpath==0.82
kafka-python==1.4.6
lxml==4.5.1
MarkupSafe==1.1.1
more-itertools==8.2.0
mysqlclient==1.3.12
nb-log==3.4
packaging==20.4
paramiko==2.7.1
pefile==2019.4.18
pika==1.1.0
pluggy==0.13.1
portend==2.6
py==1.9.0
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.9.7
PyInstaller==3.6
pymongo==3.5.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.7
pypiwin32==223
pytest==6.0.1
pytest-assume==2.2.1
pytest-forked==1.3.0
pytest-html==1.19.0
pytest-metadata==1.9.0
pytest-repeat==0.7.0
pytest-rerunfailures==9.0
pytest-xdist==1.34.0
PyTestReport==0.2.1
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
PyYAML==3.12
requests==2.23.0
requests-toolbelt==0.9.1
robotframework==3.2.1
robotframework-seleniumlibrary==4.4.0
rsa==4.0
selenium==3.141.0
six==1.14.0
tempora==3.0.0
toml==0.10.1
tomorrow3==1.1.0
urllib3==1.25.3
wcwidth==0.1.9
Werkzeug==1.0.1
WMI==1.4.9
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
zc.lockfile==2.0
zipp==3.1.0
zmail==0.2.8
